# api/docs/swagger.yaml
basePath: /
definitions:
  api.apiError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
      type:
        type: string
    type: object
  api.apiErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.apiError'
      object:
        description: '"error"'
        type: string
    type: object
...
  api.deletedResponse:
    properties:
      deleted:
        type: boolean
      id:
        type: string
      object:
        type: string
    type: object
...
  api.listResponse-api_webhookEndpointResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.webhookEndpointResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
      object:
        description: '"list"'
        type: string
    type: object
  api.listResponse-api_webhookEventResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.webhookEventResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
      object:
        description: '"list"'
        type: string
    type: object
...
  api.webhookEndpointRequest:
    properties:
      enabled:
        description: nil = do not change; false = disable; true = enable
        type: boolean
      events:
        example:
        - '["file.created"'
        - '"diff.done"]'
        items:
          type: string
        type: array
      url:
        example: https://example.com/webhook
        type: string
    type: object
  api.webhookEndpointResponse:
    properties:
      created:
        type: integer
      enabled:
        type: boolean
      events:
        items:
          type: string
        type: array
      id:
        type: string
      object:
        description: '"webhook.endpoint"'
        type: string
      url:
        type: string
    type: object
  api.webhookEventResponse:
    properties:
      attempts:
        type: integer
      created:
        type: integer
      endpoint_id:
        type: string
      id:
        type: string
      object:
        description: '"webhook.event"'
        type: string
      status:
        type: string
      type:
        type: string
    type: object
host: legist.nadevko.cc
info:
  contact: {}
  description: |-
    AI assistant for comparing revisions of normative legal acts (НПА) of the Republic of Belarus.

    ## Authentication
    All endpoints except public auth routes require a Bearer token in the `Authorization` header.
    Obtain tokens via `POST /sessions`. Refresh via `POST /tokens/refresh`.

    ## Request IDs
    Every response includes a `Request-Id` header for debugging.

    ## Idempotency
    `POST` requests require an `Idempotency-Key` header.
    Repeating a request with the same key returns the cached response.
    Keys expire after 24 hours. Reusing a key for a different endpoint returns 422.

    ## Roles
    Users have `role`: `user` or `admin` (in `users` table and JWT `role` claim).
    Only `admin` may change another user's role via `PATCH /users/:id`.
    In dev (`ENV=dev`), new registrations default to `admin`; otherwise default is `user`.

    ## Ownership and access control
    Resources belong to the authenticated user. Attempting to read or mutate
    another user's private resource returns 404 (not 403) to avoid leaking existence.
    Public resources (`user_id IS NULL`) are readable by any authenticated user but cannot be mutated.
    List `owner` query (Stripe-like): omit → non-admin sees only own rows; admin sees own + public.
    `owner=null` or `owner=public` (aliases, admin only) → public rows only. `owner=<your user id>` → own rows only. Other ids are rejected.

...

...

...

...

...

...

...

...

...

...
paths:
...
  /webhooks:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: starting_after
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listResponse-api_webhookEndpointResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: List webhook endpoints
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      parameters:
      - description: Endpoint config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.webhookEndpointRequest'
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.webhookEndpointResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Create webhook endpoint
      tags:
      - Webhooks
  /webhooks/{id}:
    delete:
      parameters:
      - description: Webhook endpoint ID
        in: path
        name: id
        required: true
        type: string
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.deletedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete webhook endpoint
      tags:
      - Webhooks
    get:
      parameters:
      - description: Webhook endpoint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.webhookEndpointResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Get webhook endpoint
      tags:
      - Webhooks
    patch:
      consumes:
      - application/json
      parameters:
      - description: Webhook endpoint ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.webhookEndpointRequest'
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.webhookEndpointResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Update webhook endpoint (url, events, enabled)
      tags:
      - Webhooks
  /webhooks/{id}/events:
    get:
      parameters:
      - description: Webhook endpoint ID
        in: path
        name: id
        required: true
        type: string
      - description: pending|delivered|failed
        in: query
        name: status
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: starting_after
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listResponse-api_webhookEventResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: List webhook delivery attempts
      tags:
      - Webhooks
securityDefinitions:
  BearerAuth:
    description: 'Enter token in format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...